+++ src/components/data/properties/object/transform/TransformScale.vue
@@ -1,36 +1,54 @@
 <template>
-	<div v-if="store.selectedObject" class="flex flex-col items-end gap-0.5">
+	<div class="flex flex-col items-end gap-0.5">
 		<InputField label="Scale X">
-			<InputNumber
-				v-model="store.selectedObject.scale.x"
-				:step="0.01"
-				:format-options="formatOptions"
-			/>
+			<InputNumber v-model="scaleX" :step="0.01" />
 		</InputField>
 		<InputField label="Y">
-			<InputNumber
-				v-model="store.selectedObject.scale.y"
-				:format-options="formatOptions"
-				:step="0.01"
-			/>
+			<InputNumber v-model="scaleY" :step="0.01" />
 		</InputField>
 		<InputField label="Z">
-			<InputNumber
-				v-model="store.selectedObject.scale.z"
-				:format-options="formatOptions"
-				:step="0.01"
-			/>
+			<InputNumber v-model="scaleZ" :step="0.01" />
 		</InputField>
 	</div>
 </template>
 
 <script lang="ts" setup>
 import { useThreeStore } from '@/store/three'
-import type { NumberFieldRootProps } from 'reka-ui'
+import { storeToRefs } from 'pinia'
+import { computed, triggerRef } from 'vue'
 
-const store = useThreeStore()
+const { selectedObject } = storeToRefs(useThreeStore())
 
-const formatOptions: NumberFieldRootProps['formatOptions'] = {
-	minimumFractionDigits: 3
-}
+const scaleX = computed({
+	get() {
+		return selectedObject.value?.scale.x
+	},
+	set(value: number) {
+		if (!selectedObject.value) return
+		selectedObject.value.scale.x = value
+		triggerRef(selectedObject)
+	}
+})
+
+const scaleY = computed({
+	get() {
+		return selectedObject.value?.scale.y
+	},
+	set(value: number) {
+		if (!selectedObject.value) return
+		selectedObject.value.scale.y = value
+		triggerRef(selectedObject)
+	}
+})
+
+const scaleZ = computed({
+	get() {
+		return selectedObject.value?.scale.z
+	},
+	set(value: number) {
+		if (!selectedObject.value) return
+		selectedObject.value.scale.z = value
+		triggerRef(selectedObject)
+	}
+})
 </script>